requirements.txt
pyTelegramBotAPI
sqlite3
python-telegram-bot==13.7